% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/can.operator.R
\name{can.operator}
\alias{can.operator}
\title{can.operator}
\usage{
can.operator(x, ...)
}
\arguments{
\item{x}{object; to test}

\item{...}{additional arguments 

\code{can.operator} test whether an object can be coerced to an operator.
Methods exist for name, function amd character classes}
}
\value{
logical
}
\description{
tests whether an object can be coerced to a operator, optionally an operator
of 'types'.
}
