% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not-in.R
\name{notin}
\alias{notin}
\alias{\%!in\%}
\title{NOT IN}
\usage{
x \%!in\% table
}
\arguments{
\item{x}{object on the lhs}

\item{table}{object/list on the rhs}
}
\description{
Commonly created NOT-IN operator
}
