\name{operators-package}
\alias{operators-package}
\docType{package}
\title{
Utilities for working with R operators
}
\description{
This package contains tools for manipulating R operators.
}
\details{
  \tabular{ll}{
    Package: \tab operator.toolss\cr
    Type: \tab Package\cr
    Version: \tab 1.0.0\cr
    Date: \tab 2010-07-18\cr
    License: \tab GPL-2\cr
  }
}

\author{
  Christopher Brown <cbrown@opendatagroup.com>
  Maintainer: Christopher Brown <cbrown@opendatagroup.com>
}

\section{Notes}{ 
  Submissions welcome.
}

\keyword{ package }

