% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeOperator.R
\name{removeOperator}
\alias{removeOperator}
\title{Unregister a an operator.}
\usage{
removeOperator(x)
}
\arguments{
\item{x}{\code{character}. The name of the operator}
}
\value{
None. Used for side-effects.
}
\description{
\code{removeOperator} unregisted an operator by removing it from the list of
operators found at \code{ .Options$operators }. All operator attributes are
that have been set will be lost.
}
\details{
Warns if the operator has not been registered.
}
\examples{

  # Unregister ? as an operator.
  \dontrun{
    removeOperator( '?' )
  }

}
\seealso{
\code{\link{setOperators}} for registering Operators.
}
\author{
Christopher Brown
}
\keyword{utilities}
