% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fun2name}
\alias{fun2name}
\alias{name2fun}
\title{Convert between a function and its name and vice versa.}
\usage{
fun2name(f)

name2fun(x)
}
\arguments{
\item{f}{function}

\item{x}{name; more specifically, an object to be converted into a name and eval'd

\code{fun2name} compares the function against existing functions using
\code{\link{identical}}. If a match is found, the name of the matching
function ( expressed as a \code{character} ) is returned.

\code{fun2name} will not work for S4 Methods.}
}
\value{
fun2name: character (name of function)
  name2fun: function
}
\description{
\code{fun2name} compares a function (body) to all defined functions. If an
identical match is found to a defined function, that function is returned.
NB. This does not search through S4 methods.
}
\details{
\code{name2fun} simply converts its argument to a name and than evals it
to produce a function definition
}

