% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/not-in.R
\name{notin}
\alias{\%!in\%}
\alias{notin}
\title{NOT IN}
\usage{
x \%!in\% table
}
\arguments{
\item{x}{object on the lhs}

\item{table}{object/list on the rhs}
}
\description{
Commonly created NOT-IN operator
}

