\name{operators-package}
\alias{operators-package}
\docType{package}
\title{
  Utilities for working with R operators
}
\description{
This package contains tools for manipulating R operators.
}
\details{
  \tabular{ll}{
    Package: \tab operator.toolss\cr
    Type: \tab Package\cr
    Version: \tab 1.3.0\cr
    Date: \tab 2012-11-30\cr
    License: \tab GPL-2\cr
  }
}

\author{
  Christopher Brown <chris.brown -at - decisionpatterns.com>
  Maintainer: Christopher Brown <chris.brown -at- decisionpatterns.com>
}

\section{Notes}{ 
  Submissions welcome.
}

\keyword{ package }

