% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/operator-tools-package.R
\docType{package}
\name{operator.tools-package}
\alias{operator.tools}
\alias{operator.tools-package}
\title{Utilities for working with R operators}
\description{
This package contains tools for manipulating R operators.
}
\details{
\tabular{ll}{ Package: \tab operator.toolss\cr Type: \tab Package\cr
Version: \tab 1.3.0\cr Date: \tab 2012-11-30\cr License: \tab GPL-2\cr }
}
\section{Notes}{
 Submissions welcome.
}
\author{
Christopher Brown <chris.brown -at - decisionpatterns.com>
Maintainer: Christopher Brown <chris.brown -at- decisionpatterns.com>
}
\keyword{package}

